Fix some Python errors

+++ plots/pi.py
@@ -103,7 +103,7 @@ graph.plot(x, y, "g+")
 
 graph = figure.add_subplot(122)
 
-N_range = logspace(0, 6, 10)
+N_range = logspace(0, 6, 10, dtype=int)
 
 graph.set_xscale("log")
 graph.set_yscale("log")
